refactor: User 엔티티 생성책임을 User로 이동 및 GlobalHandlerException을 config 패키지로 이동
