--- src/main/java/com/delivery/user/GlobalExceptionHandler.java
@@ -1,31 +1,0 @@
-package com.delivery.user;
-
-import org.springframework.dao.DuplicateKeyException;
-import org.springframework.http.ResponseEntity;
-import org.springframework.validation.FieldError;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-import java.util.LinkedHashMap;
-import java.util.Map;
-
-@RestControllerAdvice
-public class GlobalExceptionHandler {
-
-    @ExceptionHandler(MethodArgumentNotValidException.class)
-    public ResponseEntity<String> methodArgumentNotValidExceptionHandler(MethodArgumentNotValidException ex) {
-        Map<String, String> errors = new LinkedHashMap<>();
-        errors.put("errorLength", String.valueOf(ex.getFieldErrors().size()));
-        for (FieldError error : ex.getFieldErrors()) {
-            errors.put(error.getField(), error.getDefaultMessage());
-        }
-        return ResponseEntity.badRequest().body(errors.toString());
-    }
-
-
-    @ExceptionHandler(DuplicateKeyException.class)
-    public ResponseEntity<String> methodArgumentNotValidExceptionHandler(DuplicateKeyException ex) {
-        return ResponseEntity.badRequest().body("email already exists - " + ex.getMessage());
-    }
-}
